added helper.js, installed mocha and chai
