+++ package.json
@@ -27,6 +27,8 @@
     "express": "^4.17.1"
   },
   "devDependencies": {
+    "chai": "^4.3.4",
+    "mocha": "^9.1.1",
     "nodemon": "^2.0.12"
   }
 }
